vit/requirements.txt
vit-pytorch
einops
odach
timm
